requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Werkzeug==2.2.2
fpdf2==2.7.4
python-dotenv==0.21.0
